requirements.txt
scrapy>=2.5.0
scrapy-redis>=0.7.0
redis>=4.0.0
kafka-python>=2.0.2
mysql-connector-python>=8.0.26
SQLAlchemy>=1.4.23
pandas>=1.3.3
numpy>=1.21.2
requests>=2.26.0
beautifulsoup4>=4.9.3
lxml>=4.6.3
pyarrow>=5.0.0
python-dateutil>=2.8.2
dnspython>=2.1.0
prettytable>=2.2.1
psutil>=5.8.0
graphviz>=0.17
python-dotenv>=0.19.0 